# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IObenkyo api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/accounting/bills/{recordId}": {
            "delete": {
                "tags": [
                    "Accounting",
                    "Bills"
                ],
                "summary": "Allows to remove a bill",
                "operationId": "App\\Http\\Controllers\\BillController::destroy",
                "responses": {
                    "204": {
                        "description": "If the Bill was removed",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Accounting",
                    "Bills"
                ],
                "summary": "Allows to update a selected bill",
                "operationId": "App\\Http\\Controllers\\BillController::update",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "A title for the bill",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "apply_date",
                        "in": "query",
                        "description": "Date: YYYY-MM-DD HH:mm:ss",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "frequency",
                        "in": "query",
                        "description": "Id of the frequency applied to the bill",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "value",
                        "in": "query",
                        "description": "Value or price for the bill",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "If the record was updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bill"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/settings/frequencies": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Lists all frequencies",
                "operationId": "App\\Http\\Controllers\\FrequencyController::index",
                "responses": {
                    "200": {
                        "description": "List of application frequencies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Frequency"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Settings"
                ],
                "summary": "Allows to store a new frequency",
                "operationId": "App\\Http\\Controllers\\FrequencyController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Days applied for the frequency",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_daily",
                        "in": "query",
                        "description": "If the frequency applies daily",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "is_weekly",
                        "in": "query",
                        "description": "If the frequency applies every week",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "is_monthly",
                        "in": "query",
                        "description": "If the frequency applies every month",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "is_every_year",
                        "in": "query",
                        "description": "If the frequency applies every year",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "If the frequency was stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Frequency"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/settings/frequencies/{recordId}": {
            "delete": {
                "tags": [
                    "Settings"
                ],
                "summary": "Remove a frequency",
                "operationId": "App\\Http\\Controllers\\FrequencyController::destroy",
                "responses": {
                    "204": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Settings"
                ],
                "summary": "Allows to update a frequency",
                "operationId": "App\\Http\\Controllers\\FrequencyController::update",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Days applied for the frequency",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_daily",
                        "in": "query",
                        "description": "If the frequency applies daily",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "is_weekly",
                        "in": "query",
                        "description": "If the frequency applies every week",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "is_monthly",
                        "in": "query",
                        "description": "If the frequency applies every month",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "is_every_year",
                        "in": "query",
                        "description": "If the frequency applies every year",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the updated frequency",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Frequency"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/habits/categories": {
            "get": {
                "tags": [
                    "Habits"
                ],
                "summary": "Allows to list habit categories",
                "operationId": "App\\Http\\Controllers\\HabitCategoryController::index",
                "responses": {
                    "200": {
                        "description": "Categories list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/HabitCategory"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Habits"
                ],
                "summary": "Allows to save a new habit category",
                "operationId": "App\\Http\\Controllers\\HabitCategoryController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name assigned to the category",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "icon",
                        "in": "query",
                        "description": "Icon to be assinged to the category, see: https://fontawesome.com/icons?d=gallery",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Saved category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HabitCategory"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Habits"
                ],
                "summary": "Allows to remove a category",
                "operationId": "App\\Http\\Controllers\\HabitCategoryController::destroy",
                "responses": {
                    "204": {
                        "description": "If the category was removed successfuly",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Habits"
                ],
                "summary": "Allows to update a habit category",
                "operationId": "App\\Http\\Controllers\\HabitCategoryController::update",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name assigned to the category",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "icon",
                        "in": "query",
                        "description": "Icon to be assinged to the category, see: https://fontawesome.com/icons?d=gallery",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "If the category was updated successfuly",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HabitCategory"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Bill": {
                "properties": {
                    "id": {
                        "description": "",
                        "type": "integer"
                    },
                    "apply_date": {
                        "description": "Date or day of month which the bill apply",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    },
                    "custom_days": {
                        "description": "Days for the applied frequency",
                        "type": "string",
                        "default": "mon,tue,wed,thu,fri,sat,sun,*"
                    },
                    "description": {
                        "description": "A description for the bill",
                        "type": "string"
                    },
                    "is_open": {
                        "description": "If the system should keep generating this bill",
                        "type": "boolean"
                    },
                    "frequency": {
                        "$ref": "#/components/schemas/Frequency"
                    },
                    "name": {
                        "description": "A simple title for the bill",
                        "type": "string"
                    },
                    "value": {
                        "description": "Value/price, for the bill",
                        "type": "number",
                        "default": "0"
                    },
                    "created_at": {
                        "description": "Created date",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    },
                    "updated_at": {
                        "description": "Updated date",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    }
                },
                "type": "object"
            },
            "Frequency": {
                "properties": {
                    "id": {
                        "description": "",
                        "type": "integer"
                    },
                    "name": {
                        "description": "",
                        "type": "string"
                    },
                    "days": {
                        "description": "Amount of days for the frequency",
                        "type": "integer"
                    },
                    "is_daily": {
                        "description": "If apply daily",
                        "type": "boolean"
                    },
                    "is_weekly": {
                        "description": "If apply weekly",
                        "type": "boolean"
                    },
                    "is_monthly": {
                        "description": "If apply monthly",
                        "type": "boolean"
                    },
                    "is_every_year": {
                        "description": "If apply every year",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Created date",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    },
                    "updated_at": {
                        "description": "Updated date",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    }
                },
                "type": "object"
            },
...
            "HabitCategory": {
                "properties": {
                    "id": {
                        "description": "",
                        "type": "integer"
                    },
                    "name": {
                        "description": "",
                        "type": "string"
                    },
                    "icon": {
                        "description": "",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Created date",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    },
                    "updated_at": {
                        "description": "Updated date",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        []
    ]
}